refactor: 독서왕 컴포넌트 분리
